% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc1.R
\name{calc1}
\alias{calc1}
\title{Title}
\usage{
calc1(x, f)
}
\arguments{
\item{x}{input vector}

\item{f}{function to be applied to x}
}
\value{
result of f(x)
}
\description{
Title
}
\examples{
calc1(1:130, mean)
}
