\name{extraTrees}
\alias{extraTrees}
\title{Function to say hello from a Java class}
\description{
  This function executes a method of a Java class
  defined in a JAR inside the R package.
}
\usage{
  extraTrees()
}
\details{
  Details are given in the package vignette.
}
\value{
  The string "Hello Java World!" is returned.
}
\author{extraTrees}
\examples{
  extraTrees()
}
\keyword{manip}
